Infer method-wise type arguments
for method calls other than generic_new too
